Added transmembrane_diffusion model for secretion
New model for secretion that is an alternative for the (secretion rate, uptake rate, saturation density) model before.
Export rate is unaffected.
The model assumes that the intracellular and extracellular concentrations converge to the same value while conserving the mass of the substrate between the two volumes.
Simplifying assumptions are made analogous to the defualt model: a cell only interacts with the volume of the voxel it is in, the cell does not exclude volume in the voxel for calculating volume of the voxel, export can be split off from the rest accurately enough on this short timescale.
This new model is solved analytically on the timestep [0, diffusion_dt] (split between secretion and export as mentioned above).
Since the "default" model used BLAS-style operations to update all substrates simultaneously, I went with the simple approach and said all secretions follow the "default" model or the new "transmembrane_diffusion" model for each cell_definition.

To implement: add attribute "model" to a secretion element in the configuration file. See template project for example.

allow for model attr omitted, remove print